rename modules to pyrite
